refactor: remove the extra home page, render home in index.js and actual app in frame.js
