+++ src/pages/frame.jsx
@@ -1,11 +1,19 @@
-import React from 'react';
+import React, { useState, useRef } from 'react';
+
+// Libraries
+import useImage from 'use-image';
 import styled from 'styled-components';
 import tw from 'twin.macro';
+
+// Components
 import Carousel from '../components/frameControllers/carousel';
 import Head from '../components/shared/Head';
 import Inputs from '../components/frameControllers/input';
 import Download from '../components/toolbox/Download';
-import Canvas  from '../components/canvas/stage';
+import Canvas from '../components/canvas/Stage';
+
+// Assets
+import frameData from '../../config/frameData';
 
 const Container = styled.div`
   ${tw`
@@ -22,44 +30,47 @@ const Container = styled.div`
     `}
 `;
 
-function Frame({ 
-  stageRef, 
-  userName, 
-  setUserName, 
-  guildName, 
-  setGuildName,
-  frameImg,
-  image,
-  setSelectedFrame,
-  uploadedImage,
-  setUploadedImage,
-  FRAMES }) {
+const FRAMES = {
+  ONE: frameData.frames.ONE,
+  TWO: frameData.frames.ONE,
+  THREE: frameData.frames.TWO,
+  FOUR: frameData.frames.THREE,
+  FIVE: frameData.frames.FOUR,
+  SIX: frameData.frames.FIVE,
+};
 
-   return(
+export default function Home() {
+  const [selectedFrame, setSelectedFrame] = useState(FRAMES.ONE);
+  const [uploadedImage, setUploadedImage] = useState();
+  const [userName, setUserName] = useState('Your Name');
+  const [guildName, setGuildName] = useState('Guild Name');
+
+  const stageRef = useRef(null);
+  const [frameImg] = useImage(selectedFrame, 'Anonymous');
+  const [image] = useImage(uploadedImage, 'Anonymous');
+
+  return (
     <Container>
       <Head />
       <div className='App' style={{ display: 'grid', margin: '20px' }}>
         <Canvas
-         stageRef={stageRef}
-         userName={userName}
-         guildName={guildName}
-         frameImg={frameImg}
-         image={image}
-        />       
+          stageRef={stageRef}
+          userName={userName}
+          guildName={guildName}
+          frameImg={frameImg}
+          image={image}
+        />
         <Carousel frames={FRAMES} setSelectedFrame={setSelectedFrame} />
         <Inputs
-         uploadedImage={uploadedImage}
-         setUploadedImage={setUploadedImage}
-         userName={userName}
-         setUsername={setUserName}
-         guildName={guildName}
-         setGuildname={setGuildName}
-         />
-
+          uploadedImage={uploadedImage}
+          setUploadedImage={setUploadedImage}
+          userName={userName}
+          setUsername={setUserName}
+          guildName={guildName}
+          setGuildname={setGuildName}
+        />
         <Download stageRef={stageRef} />
       </div>
     </Container>
-   );
-};
-
-export default Frame;
+  );
+}
